run reset command to remove example UI to begin frontend dev
