+++ frontend/app-example/(tabs)/_layout.tsx
